requirements.txt
flask
groq
firebase-admin
python-dotenv
requests
